+++ components/PlanAuthoring/demo/demoData.ts
@@ -16,7 +16,7 @@ import {
   PlanQuestionDefinition,
   PlanSectionDefinition,
   computeProgress,
-} from "./model";
+} from "../model";
 
 // Guidance bodies are WYSIWYG-authored in the admin, so demo data deliberately
 // mixes headings, links, bold/italic, lists, and blockquotes at varying lengths.
@@ -1013,132 +1013,5 @@ export function createPlanAuthoringDemo(): PlanAuthoringModel {
     sections,
     availableGuidanceOrgs,
     selectedGuidanceOrgIds: ["nih", "uci", "stanford"],
-  };
-}
-
-export function createLabQuestionDemo(
-  overrides: Partial<PlanQuestionDefinition> = {}
-): PlanQuestionDefinition {
-  const base = createPlanAuthoringDemo().sections[0].questions[0];
-  return {
-    ...base,
-    ...overrides,
-    identity: overrides.identity ?? base.identity,
-    sectionIdentity: overrides.sectionIdentity ?? base.sectionIdentity,
-    parsedJson: overrides.parsedJson ?? base.parsedJson,
-    guidanceSources: overrides.guidanceSources ?? base.guidanceSources,
-    comments: overrides.comments ?? base.comments,
-  };
-}
-
-export function createStressPlanAuthoringDemo(
-  options: { sectionCount?: number; questionsPerSection?: number } | number = {},
-  questionsPerSectionArg?: number
-): PlanAuthoringModel {
-  const sectionCount =
-    typeof options === "number" ? options : (options.sectionCount ?? 8);
-  const questionsPerSection =
-    typeof options === "number"
-      ? (questionsPerSectionArg ?? 12)
-      : (options.questionsPerSection ?? 12);
-  const seed = createPlanAuthoringDemo();
-  const questionTypeCycle = [
-    TEXT_AREA_QUESTION_TYPE,
-    TEXT_FIELD_QUESTION_TYPE,
-    BOOLEAN_QUESTION_TYPE,
-    RADIOBUTTONS_QUESTION_TYPE,
-    CHECKBOXES_QUESTION_TYPE,
-    SELECTBOX_QUESTION_TYPE,
-    NUMBER_QUESTION_TYPE,
-    DATE_QUESTION_TYPE,
-  ];
-
-  const sections: PlanSectionDefinition[] = Array.from(
-    { length: sectionCount },
-    (_, sectionIndex) => {
-      const isCustom = sectionIndex % 5 === 4;
-      const sectionIdentity = isCustom
-        ? { kind: "custom" as const, customSectionId: 1000 + sectionIndex }
-        : { kind: "base" as const, versionedSectionId: 1000 + sectionIndex };
-
-      const questions: PlanQuestionDefinition[] = Array.from(
-        { length: questionsPerSection },
-        (_, questionIndex) => {
-          const questionType =
-            questionTypeCycle[questionIndex % questionTypeCycle.length];
-          const answered = (sectionIndex + questionIndex) % 3 !== 0;
-          const identity = isCustom
-            ? {
-                kind: "custom" as const,
-                customQuestionId: 10000 + sectionIndex * 100 + questionIndex,
-              }
-            : {
-                kind: "base" as const,
-                versionedQuestionId: 10000 + sectionIndex * 100 + questionIndex,
-              };
-
-          return makeQuestion({
-            identity,
-            sectionIdentity,
-            title: `Section ${sectionIndex + 1} · Question ${questionIndex + 1} (${questionType})`,
-            requirementHtml:
-              questionIndex % 4 === 0
-                ? "<p>Funder expects a concise, concrete response.</p>"
-                : undefined,
-            required: questionIndex % 2 === 0,
-            questionType,
-            parsedJson: {
-              type: questionType,
-              options:
-                questionType === RADIOBUTTONS_QUESTION_TYPE ||
-                questionType === CHECKBOXES_QUESTION_TYPE ||
-                questionType === SELECTBOX_QUESTION_TYPE
-                  ? [
-                      { label: "Option A", value: "a" },
-                      { label: "Option B", value: "b" },
-                      { label: "Option C", value: "c" },
-                    ]
-                  : undefined,
-            },
-            answerJson: answered
-              ? {
-                  type: questionType,
-                  answer:
-                    questionType === BOOLEAN_QUESTION_TYPE
-                      ? true
-                      : questionType === CHECKBOXES_QUESTION_TYPE
-                        ? ["a"]
-                        : questionType === NUMBER_QUESTION_TYPE
-                          ? 42
-                          : questionType === TEXT_AREA_QUESTION_TYPE
-                            ? "<p>Sample stress-test answer.</p>"
-                            : "Sample answer",
-                }
-              : null,
-            hasAnswer: answered,
-            comments: answered && questionIndex === 0 ? sampleComments : [],
-            displayOrder: questionIndex + 1,
-          });
-        }
-      );
-
-      return {
-        identity: sectionIdentity,
-        title:
-          sectionIndex < seed.sections.length
-            ? `${seed.sections[sectionIndex % seed.sections.length].title} (${sectionIndex + 1})`
-            : `Generated section ${sectionIndex + 1}`,
-        introductionHtml: `<p>Stress demo section ${sectionIndex + 1}.</p>`,
-        displayOrder: sectionIndex + 1,
-        questions,
-      };
-    }
-  );
-
-  return {
-    ...seed,
-    title: `${seed.title} · stress demo`,
-    progress: computeProgress(sections),
-    sections,
   };
 }
