+++ components/PlanAuthoring/demo/index.ts
@@ -1,0 +1,5 @@
+export { createPlanAuthoringDemo } from "./demoData";
+export {
+  createPlanAuthoringDemoDataSource,
+  type DemoDataSourceOptions,
+} from "./demoDataSource";
