+++ components/PlanAuthoring/index.ts
@@ -1,4 +1,4 @@
-import {
+import type {
   PlanAuthoringModel,
   PlanCapabilities,
   PlanComment,
@@ -10,21 +10,8 @@ import {
   PlanQuestionSaveState,
   PlanSectionDefinition,
   PlanSectionIdentity,
-  computeProgress,
-  questionAnchorId,
-  questionKey,
-  sectionAnchorId,
-  sectionKey,
 } from "./model";
-import {
-  createLabQuestionDemo,
-  createPlanAuthoringDemo,
-  createStressPlanAuthoringDemo,
-} from "./demoData";
-import {
-  PlanAuthoringDataSource,
-  createPlanAuthoringDemoDataSource,
-} from "./dataSource";
+import type { PlanAuthoringDataSource } from "./dataSource";
 
 export type {
   PlanAuthoringModel,
@@ -41,27 +28,4 @@ export type {
   PlanAuthoringDataSource,
 };
 
-export {
-  computeProgress,
-  questionAnchorId,
-  questionKey,
-  sectionAnchorId,
-  sectionKey,
-  createLabQuestionDemo,
-  createPlanAuthoringDemo,
-  createStressPlanAuthoringDemo,
-  createPlanAuthoringDemoDataSource,
-};
-
 export { default as PlanAuthoring } from "./PlanAuthoringScreen";
-export { default as PlanSectionNavigation } from "./PlanSectionNavigation";
-export { default as PlanSectionPickerDialog } from "./PlanSectionPickerDialog";
-export { default as PlanSection } from "./PlanSection";
-export { default as PlanQuestion } from "./PlanQuestion";
-export { default as PlanQuestionHeader } from "./PlanQuestionHeader";
-export { default as PlanQuestionAnswer } from "./PlanQuestionAnswer";
-export { default as PlanQuestionSaveStatus } from "./PlanQuestionSaveStatus";
-export { default as PlanQuestionSidebar } from "./PlanQuestionSidebar";
-export { default as PlanGuidanceTabs } from "./PlanGuidanceTabs";
-export { default as PlanGuidanceCustomizeDialog } from "./PlanGuidanceCustomizeDialog";
-export { default as PlanComments } from "./PlanComments";
